Moved imgui-skia-renderer to it's own repo and cleaned up main.
Fixed ctrl+Q.
Refactored main.
Moved imgui to it's own repo.
